fixed size of ledger height and gave it scrollbars
now the ledger itself takes up one screen. There will be double scrollbars
(one for the leger, one for the browser) once other widgets are added, and
I;m not sure if that's a good idea. The alternative it the date range
dictates the length of the page, possibly pushing widgets above and below
the ledger far apart. I will have to make more widgets before this can be
determined.

For now it works, as the range selector sub-widget stays in place.