new columns in ledger entered and queried
The new columns are documented in the mysql_notes, and the Ledger query
has been modified to only display items where the DEL column is NULL.
Deleted entries are left in the source table, just not queried for the
ledger. This way they can be recoverable by altering this flag value.