getSource to check ledger date change
get source added to change in calendar date for a ledger entry. Right now,
scaffolding announces the returned value.
This value should be checked, if source is download, continue with existing
handling of a date change. But if not, then just update the date value in the
entry in the same way we do a comment or the serial.