added script tags for dialog
Will need scripts for modal dialog for the import csv functions, so
scripts added.

also streamlined the script drawing to a loop, since jquery-ui requires so
many script additions, I now just draw them to the page in a loop.

To add a new script, I just add its path to the aray.