Dialog to upload CSV files for import
Clicking the import button opens a modal dialog box. The box contains form
elements for uploading the csv file to the server, where php scripts will
(in the near future) import the data into sql.

In order for this to work, you must have your php.ini configured for
uploads. See the PHP manual for instructions on making this work.

right now, the file selected is simply uploaded to the tmp directory and
nothing is done. Will need to add processing. I will pass the file through
a dummy validation function. In case anyone that is not me attempts to
use moola, validation scripts will need to be developed.

But that's boring. I'll get back to it when it's necessary.