more work on loading the csv
I under-thought the process of getting the filename to the import script.
I'm at risk of creating an overly convoluted way of doing it, that is more
complicated than it needs to be if I continue throwing code at the
problem.

So I'm just gonna commit this, push it, and think about it for a few days
in my spare brain-cycles.

The answer is in there, I just need to put it on paper first.