Added Category module
Needs functionality. Right now it's a text and a button that does nothing on a
hook. Button will insert to a category table. This table will allow assigning
categories to ledger entries.

Obviously, I got bored on the import integration. It's on the sidelines for now
while I cut to the core functions I want this to perform. It will have to be
command line imports until I come back to it.